password manager/requirements.txt
requests
pyperclip
pycryptodomex
flask
flask-cors